searchVectorDB/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
google-cloud-storage==2.9.0
google-cloud-aiplatform==1.35.0
protobuf==4.24.4